backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pydantic 
python-multipart
python-jose 
passlib==1.7.4 
bcrypt==4.0.1
pymongo[srv]

python-dotenv
motor
ogl